feat: created a MealList component to be used on the CategoryMealScreen and FavoritesScreen
