+++ screens/CategoryMealScreen.js
@@ -9,6 +9,7 @@ import {
 
 import  { CATEGORIES, MEALS } from '../data/dummy-data';
 import MealItem from '../components/MealItem';
+import MealList from '../components/MealList';
 
 const CategoryMealScreen = props =>{
    const catId = props.navigation.getParam('categoryId');
@@ -16,34 +17,11 @@ const CategoryMealScreen = props =>{
        meal => meal.categoryIds.includes(catId));
   // he did meal.categoryId.indexOf(catId) >=0
 
-
-    const renderItemMeal = (itemData)=>{
-        return <MealItem 
-                    title={itemData.item.title} 
-                    duration={itemData.item.duration} 
-                    complexity={itemData.item.complexity}
-                    affordability={itemData.item.affordability}
-                    image={itemData.item.imageURL}
-                    onSelectMeal={()=>{
-                        props.navigation.navigate(
-                            { routeName:'MealDetail', 
-                              params:{ mealId: itemData.item.id }
-                    })    
-                }}
-                />
-    };
- 
-    return (
-        <View style={styles.screen}>
-            <FlatList 
-                 keyExtractor={(item, index)=>{item.id}}
-                 data={displayedMeals}
-                 renderItem={renderItemMeal}
-                 showVerticalScrollIndicator={false}
-                 style={{width: "96%", marginVertical: 8}}
-            />
-        </View>
-    );
+    return <MealList 
+                listData={displayedMeals}
+                navigation={props.navigation}
+            
+            />;
 };
 
 CategoryMealScreen.navigationOptions = (navigationData)=>{
@@ -57,14 +35,5 @@ CategoryMealScreen.navigationOptions = (navigationData)=>{
     return  {  headerTitle: selectedCategory.title }
 };
    
-const styles =  StyleSheet.create({
-    screen:{
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center'
-    }, 
-});
-
-
 
 export default CategoryMealScreen;
